+++ src/main/java/net/meander/subtlyd/mixin/client/renderer/extract/LevelExtractorMixin.java
@@ -20,19 +20,21 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(LevelExtractor.class)
 public class LevelExtractorMixin {
     @Shadow private ClientLevel level;
+    private final Minecraft minecraft = Minecraft.getInstance();
 
     @Inject(method = "isEntityVisible", at = @At("RETURN"), cancellable = true)
-    private void cullEntityBehindBlocks(final Entity entity, final Frustum frustum, final double camX, final double camY, final double camZ, CallbackInfoReturnable<Boolean> cir) {
-        Options options = Minecraft.getInstance().options;
+    private void cullOccludedEntity(final Entity entity, final Frustum frustum, final double camX, final double camY, final double camZ, CallbackInfoReturnable<Boolean> cir) {
+        Options options = minecraft.options;
 
         if (options.entityCulling().get() == EntityCullingMethod.OCCLUSION) {
             if (cir.getReturnValue() && level != null) {
                 Vec3 cameraPos = new Vec3(camX, camY, camZ);
+                double distanceSqr = cameraPos.distanceToSqr(entity.position());
 
-                if (cameraPos.distanceToSqr(entity.position()) <= options.entityDistanceScaling().get() * 25600.0) {
-                    if (OcclusionManager.getInstance().isEntityOccluded(entity, cameraPos, level)) {
-                        cir.setReturnValue(false);
-                    }
+                if (distanceSqr > OcclusionManager.getInstance().getCurrentFogEndSqr()) {
+                    cir.setReturnValue(false);
+                } else if (OcclusionManager.getInstance().isEntityOccluded(entity, cameraPos, level)) {
+                    cir.setReturnValue(false);
                 }
             }
         }